# dist/build/swagger.json
{
    "components": {
        "examples": {},
        "headers": {},
        "parameters": {},
        "requestBodies": {},
        "responses": {},
        "schemas": {
            "TestapiResponse": {
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "expiresIn": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "required": [
                    "token",
                    "expiresIn"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "MsException": {
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "details": {}
                },
                "required": [
                    "code",
                    "message"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "ServiceResult_TestapiResponse_": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/TestapiResponse"
                    },
                    "error": {
                        "$ref": "#/components/schemas/MsException"
                    }
                },
                "type": "object",
                "additionalProperties": false
            }
        },
        "securitySchemes": {}
    },
    "info": {
        "title": "webtoapp",
        "version": "1.0.0",
        "license": {
            "name": "MIT"
        },
        "contact": {}
    },
    "openapi": "3.0.0",
    "paths": {
        "/testapi/{userId}": {
            "get": {
                "operationId": "GetUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServiceResult_TestapiResponse_"
                                }
                            }
                        }
                    }
                },
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "format": "double",
                            "type": "number"
                        }
                    },
                    {
                        "in": "query",
                        "name": "name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/testapi": {
            "post": {
                "operationId": "CreateTestapi",
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServiceResult_TestapiResponse_"
                                }
                            }
                        }
                    }
                },
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "test": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "test"
                                ],
                                "type": "object"
                            }
                        }
                    }
                }
            }
        }
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}